requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
Cython==3.0.11
fake-bpy-module-latest==20240826
fastprogress==1.0.3
idna==3.8
mathutils==3.3.0
numpy==2.1.0
requests==2.32.3
urllib3==2.2.2
zstandard==0.23.0